requirements.txt
numpy
pandas
tqdm
faiss-cpu
pickle-mixin
ollama
requests
nltk
scikit-learn
sentence-transformers
transformers
torch
